tour_api/requirements.txt
asgiref==3.9.1
certifi==2025.7.14
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.3.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
